+++ DangerMoleMobile/app/src/main/java/com/example/dangermolemobile/Utility.kt
@@ -8,6 +8,8 @@ import android.content.pm.PackageManager
 import android.support.v4.content.ContextCompat
 import android.app.Activity
 import android.support.v4.app.ActivityCompat
+import java.math.RoundingMode
+import java.text.DecimalFormat
 
 class Utility {
     private val REQUEST_CAM_STORAGE_PERMISSIONS_CODE = 298
@@ -30,6 +32,15 @@ class Utility {
                 Manifest.permission.READ_EXTERNAL_STORAGE, Manifest.permission.CAMERA), REQUEST_CAM_STORAGE_PERMISSIONS_CODE )
         }
 
+    }
+
+    fun floatSanitizer(float: Float): String{
+        var number = float
+        number = (number * 100)
+        val dec = DecimalFormat("##.##")
+        dec.roundingMode = RoundingMode.CEILING
+        number = dec.format(number).toFloat()
+        return number.toString()
     }
 
 }
